+++ CHANGES.md
@@ -5,9 +5,12 @@ Changes in v1.2rc1
 ------------------
 
 - Added explicit support for running macOS printer applications as a server.
+- Added unit test support for the new SNMP-based supply level and status
+  monitoring code.
 - Fixed an issue with the default system callback for `papplMainloop`.
 - Fixed an issue with `papplDeviceList` and DNS-SD discovery when there was no
   active system.
+- Fixed printer compatibility issues with the new `papplDeviceGetSupplies` API.
 
 
 Changes in v1.2b1